added more infos servers
started to create differents files with differents fonctions. i added a function to get the number of player connected during a time given and im also creating a function to get the basic infos of the server. the function is now returning a dataframe and a csv file. to do : returning a json once the data is clean